backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
tensorflow==2.19.0
pillow
numpy==2.1.3
python-dotenv
supabase
werkzeug
logging
google-cloud-aiplatform
google-cloud-core
google-api-core
google-genai
googleapis-common-protos
uuid